[style] 드랍다운 아이콘 수정 및 border props 추가 #6
