+++ src/components/Abstraction/DropDown/DropDownList.module.scss
@@ -1,0 +1,15 @@
+.dropDownList {
+  cursor: pointer;
+  border: 0.1rem solid #abb0b7;
+  border-bottom: none;
+  padding: 0.4rem 2rem 0.4rem 0.8rem;
+  &:first-child {
+    &.borderTop {
+      border-top: none;
+    }
+  }
+
+  &:last-child {
+    border-bottom: 0.1rem solid #abb0b7;
+  }
+}
