--- src/components/Abstraction/DropDownList.module.scss
@@ -1,6 +1,0 @@
-.dropDownList {
-  display: flex;
-  align-items: center;
-  gap: 0.8rem;
-  justify-content: space-between;
-}
